Create LearningResource and LearningRecord utilities.
