+++ src/test/java/com/deloitte/elrr/aggregator/test/consumer/ProcessCompletedTest.java
@@ -15,6 +15,8 @@ import com.deloitte.elrr.aggregator.consumer.ProcessPerson;
 import com.deloitte.elrr.aggregator.rules.ProcessCompleted;
 import com.deloitte.elrr.aggregator.test.util.TestFileUtils;
 import com.deloitte.elrr.aggregator.utils.LangMapUtil;
+import com.deloitte.elrr.aggregator.utils.LearningRecordUtil;
+import com.deloitte.elrr.aggregator.utils.LearningResourceUtil;
 import com.deloitte.elrr.entity.Person;
 import com.deloitte.elrr.jpa.svc.EmailSvc;
 import com.deloitte.elrr.jpa.svc.IdentitySvc;
@@ -41,6 +43,10 @@ class ProcessCompletedTest {
   @Mock LearningResourceSvc learningResourceSvc;
 
   @Mock LearningRecordSvc learningRecordSvc;
+  
+  @Mock LearningResourceUtil learningResourceUtil;
+
+  @Mock LearningRecordUtil learningRecordUtil;
 
   @InjectMocks ProcessPerson processPerson;
